feat: usePriceRange
